pyproject.toml
[tool.poetry]
name = "django-studybudy-yt"
version = "0.1.0"
description = "A threads app with Django"
authors = ["Shahryar Tayeb <shahryar.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
markdown = "^3.5.1"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
virtualenv = "^20.24.6"
django-allauth = "^0.58.2"
django-soft-delete = "^0.9.21"
django-easy-audit = "^1.3.5"
mysqlclient = "^2.2.0"
django-invitations = "^2.0.0"
django-htmx = "^1.17.0"
django-mdeditor = "^0.1.20"
django-vite-plugin = "^1.3.3"
django-guardian = "^2.4.0"
gunicorn = "^21.2.0"
environ = "^1.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
pydotplus = "^2.0.2"
nplusone = "^1.0.0"
factory-boy = "^3.3.0"
whitenoise = "^6.6.0"
halo = "^0.0.31"
pre-commit = "^3.5.0"

[tool.ruff]
line-length=120
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "**/migrations",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
